move example types within files
