+++ spacy_llm/tasks/util/__init__.py
@@ -1,6 +1,4 @@
-from .examples import SpanExample
 from .parsing import find_substrings
 from .serialization import SerializableTask
-from .span import SpanTask
 
-__all__ = ["SpanExample", "find_substrings", "SpanTask", "SerializableTask"]
+__all__ = ["find_substrings", "SerializableTask"]
